Move modules into organisms
