+++ src/components/Organisms/CookieBanner/CookieBanner.test.js
