+++ src/index.js
@@ -45,18 +45,18 @@ export { default as Card } from './components/Molecules/Card/Card';
 export {
   default as ArticleTeaser
 } from './components/Molecules/ArticleTeaser/ArticleTeaser';
-export { default as Header } from './components/Molecules/Header/Header';
+export { default as Header } from './components/Organisms/Header/Header';
 export {
   default as CookieBanner
-} from './components/Molecules/CookieBanner/CookieBanner';
-export { default as Donate } from './components/Molecules/Donate/Donate';
+} from './components/Organisms/CookieBanner/CookieBanner';
+export { default as Donate } from './components/Organisms/Donate/Donate';
 export {
   default as DoubleCopy
 } from './components/Molecules/DoubleCopy/DoubleCopy';
 export {
   default as PartnerLink
 } from './components/Molecules/PartnerLink/PartnerLink';
-export { default as Footer } from './components/Molecules/Footer/Footer';
+export { default as Footer } from './components/Organisms/Footer/Footer';
 export {
   default as SearchResult
 } from './components/Molecules/SearchResult/SearchResult';
@@ -65,13 +65,13 @@ export {
 } from './components/Molecules/SearchInput/SearchInput';
 export {
   default as EmailSignUp
-} from './components/Molecules/EmailSignUp/EmailSignUp';
+} from './components/Organisms/EmailSignUp/EmailSignUp';
 export {
   default as ShareButton
 } from './components/Molecules/ShareButton/ShareButton';
 export {
   default as Membership
-} from './components/Molecules/Membership/Membership';
+} from './components/Organisms/Membership/Membership';
 export {
   default as VideoBanner
 } from './components/Molecules/VideoBanner/VideoBanner';
